ada: Fix spurious Constraint_Error raised by 'Value of fixed-point types
This happens for very large Smalls with regard to the size of the mantissa,
because the prerequisites of the implementation used in this case are not
met, although they are documented in the head comment of Integer_To_Fixed.

This change documents them at the beginning of the body of System.Value_F
and adjusts the compiler interface accordingly.

gcc/ada/ChangeLog:

	* libgnat/s-valuef.adb: Document the prerequisites more precisely.
	* libgnat/a-tifiio.adb (OK_Get_32): Adjust to the prerequisites.
	(OK_Get_64): Likewise.
	* libgnat/a-tifiio__128.adb (OK_Get_32): Likewise.
	(OK_Get_64): Likewise.
	(OK_Get_128): Likewise.
	* libgnat/a-wtfiio.adb (OK_Get_32): Likewise.
	(OK_Get_64): Likewise.
	* libgnat/a-wtfiio__128.adb (OK_Get_32): Likewise.
	(OK_Get_64): Likewise.
	(OK_Get_128): Likewise.
	* libgnat/a-ztfiio.adb (OK_Get_32): Likewise.
	(OK_Get_64): Likewise.
	* libgnat/a-ztfiio__128.adb (OK_Get_32): Likewise.
	(OK_Get_64): Likewise.
	(OK_Get_128): Likewise.
	* exp_imgv.adb (Expand_Value_Attribute): Adjust the conditions under
	which the RE_Value_Fixed{32,64,128} routines are called for ordinary
	fixed-point types.