ada: Incorrect resolution of prefixed calls with overriding private subprogram
The compiler incorrectly treats an overriding private subprogram that
should not be visible outside a package (because it only overrides in
the private part) as a possible interpretation for a call using prefixed
notation outside of the package. This can result in an ambiguity if there
is another subprogram with the same name but a different profile declared
in the visible part of the package, or can result in resolving to the
private operation in cases where it shouldn't resolve. This happens due
to the compiler improperly concluding that the private overriding subprogram
overrides an inherited subprogram in the package visible part, even though
the only inherited subprogram is in the private part, as a result of
a misuse of the Overridden_Operation field, which, contrary to what
its name suggests, actually refers to operations of the parent type,
rather than to the operations derived from the parent's operations.

gcc/ada/ChangeLog:

	* einfo.ads: Document new field Overridden_Inherited_Operation and
	list it as a field for the entity kinds that it applies to.
	* gen_il-fields.ads (type Opt_Field_Enum): Add new literal
	Overridden_Inherited_Operation to the type.
	* gen_il-gen-gen_entities.adb: Add Overridden_Inherited_Operation as
	a field of entities of kinds E_Enumeration_Literal and Subprogram_Kind.
	* sem_ch4.adb (Is_Callable_Private_Overriding): Change name (was
	Is_Private_Overriding). Replace Is_Hidden test on Overridden_Operation
	with test of Is_Hidden on the new field Overridden_Inherited_Operation.
	* sem_ch6.adb (New_Overloaded_Entity): Set the new field
	Overridden_Inherited_Operation on an operation derived from
	an interface to refer to the inherited operation of a private
	extension that's overridden by the derived operation. Also set
	that field in the more common cases of an explicit subprogram
	that overrides, to refer to the inherited subprogram that is
	overridden. (Contrary to its name, the Overridden_Operation
	field of the overriding subprogram, which is also set in these
	places, refers to the *parent* subprogram from which the inherited
	subprogram is derived.) Also, remove a redundant Present (Alias (S))
	test in an if_statement and the dead "else" part of that statement.