ada: Missing component clause warning for discriminant of Unchecked_Union type
Even when -gnatw.c is enabled, no warning about a missing component clause
should be generated if the placement of a discriminant of an Unchecked_Union
type is left unspecified in a record representation clause (such a discriminant
occupies no storage). In determining whether to generate such a warning, in
some cases the compiler would incorrectly ignore an Unchecked_Union pragma
occurring after the record representation clause. This could result in a
spurious warning.

gcc/ada/ChangeLog:

	* sem_ch13.adb (Analyze_Record_Representation_Clause): In deciding
	whether to generate a warning about a missing component clause, in
	addition to calling Is_Unchecked_Union also call a new local
	function, Unchecked_Union_Pragma_Pending, which checks for the
	case of a not-yet-analyzed Unchecked_Union pragma occurring later
	in the declaration list.