ada: Do not inline function returning on the secondary stack
When inlining function calls that return on the secondary stack used as
function actual or in a return statement, the compiler creates an
invalid GNAT Tree with a variable of an unconstrained type without an
initializer.

Also add an extra assertion to catch problematic cases directly in
Expand_Inlined_Call.

gcc/ada/ChangeLog:

	* exp_ch6.adb (Convert): Do not call Expand_Inlined_Call for
	unsupported cases.
	* inline.adb (Expand_Inlined_Call): Add assert to catch unsupported
	case.

Co-authored-by: Eric Botcazou <botcazou@adacore.com>