ada: Fix unnecessary extra RE_Activation_Chain_Access with No_Task_Parts
This patch checks the presence of No_Task_Parts on any ancestor or
inherited interface, not only its root type, since No_Task_Parts
prohibits tasking for any of its descendant. In case the current
subprogram is overridden/inherited, we need to return the same value
we would return for the original corresponding operation. The aspect
No_Task_Parts is nonoverridable and applies also when specified in a
partial view.

gcc/ada/ChangeLog:

	* sem_ch6.adb (Might_Need_BIP_Task_Actuals): Check whether No_Task_Parts is enabled in any
	of the derived types, or interfaces, from the user-defined primitive return type.
	* sem_ch13.adb (Analyze_Aspect_Specifications): Add No_Task_Parts and No_Controlled_Parts to
	the representation chain to be visible in the full view of private types.
	* aspects.ads (Nonoverridable_Aspect_Id): As per GNAT RM, No_Task_Parts is nonoverridable.
	* sem_util.adb (Check_Inherited_Nonoverridable_Aspects): Likewise.
	* sem_util.ads: Fix typo and style.
	* sem_disp.adb: Missing comment.