ada: Fix generation of Initialize and Adjust calls
Before this patch, Make_Init_Call and Make_Adjust_Call made the
assumption that if the type they were called with was untagged and a
derived type, it was the untagged private view of a tagged type. That
assumption made it possible to inspect the root type's primitives to
handle the case where the underlying type was implicitly generated by
the compiler without all inherited primitives.

The introduction of the Finalizable aspect broke that assumption, so
this patch adds a new field to type entities that make the generated
full view stand out, and updates Make_Init_Call and Make_Adjust_Call to
only jump to the root type when they're passed one of those generated
types.

Make_Final_Call and Finalize_Address are two other subprograms that
perform the same test on the types they're passed. They did not suffer
from the same bug as Make_Init_Call and Make_Adjust_Call because of an
earlier, more ad hoc fix, but this patch switches them over to the newly
introduced mechanism for the sake of consistency.

gcc/ada/ChangeLog:

	* gen_il-fields.ads (Is_Implicit_Full_View): New field.
	* gen_il-gen-gen_entities.adb (Type_Kind): Use new field.
	* einfo.ads (Is_Implicit_Full_View): Document new field.
	* exp_ch7.adb (Make_Adjust_Call, Make_Init_Call, Make_Final_Call): Use
	new field.
	* exp_util.adb (Finalize_Address): Likewise.
	* sem_ch3.adb (Copy_And_Build): Set new field.