ada: Implement accessibility legality check for discriminated function result.
If a function result type has an access discriminant, then we already
generate a run-time accessibility check for a return statement. But if
we know statically that the check (if executed) is going to fail, then
that should be rejected at compile-time as a violation of RM 6.5(5.9).
Add this additional compile-time check.

gcc/ada/ChangeLog:

	* exp_ch6.adb (Apply_Access_Discrims_Accessibility_Check): If the
	accessibility level being checked is known statically, then
	statically check it against the level of the function being
	returned from.