ada: Platform-specific import for read() and write()
read() and write() return int on windows, whereas
on Posix systems the return type is ssize_t (effectively long_int).

The object file was added to GNAT_ADA_OBJS only, although the unit is also
in the compilation closure of GNATbind, but this was harmless because the
object file was essentially empty; that is no longer the case.

gcc/ada/ChangeLog:

	* libgnat/s-crtl.ads (read, write): remove import
	* libgnat/s-crtl__mingw.adb: body for windows
	* libgnat/s-crtl.adb: body for all the other targets
	* Makefile.rtl: configure s-crtl.adb/libgnat/s-crtl__mingw.adb
	* gcc-interface/Make-lang.in (GNAT_ADA_OBJS): Alphabetize.
	(GNATBIND_OBJS): Add ada/libgnat/s-crtl.o.

Co-authored-by: Eric Botcazou <ebotcazou@adacore.com>