ada: Remove unnecessary if statement
Calling Check_Nonoverridable_Aspects only makes sense when the full type
declaration being analyzed is the completion of a partial view, and the
one call site of this procedure ensures this. Therefore the handling of
all the possible cases of completion in the procedure that this patch
removes was useless.

gcc/ada/ChangeLog:

	* sem_ch3.adb (Check_Nonoverridable_Aspects): Remove if statement.