ada: Set Related_Expression on compiler-generated Valid_Scalars functions
When creating the local functions to implement the Valid_Scalars
attribute for array and record types, set a Related_Expression that
points to the original attribute reference (blah'Valid_Scalars).
This allows the Inspector to give a more user-friendly name
when these functions are called and they are known, for example,
to always return True.

gcc/ada/ChangeLog:

	* exp_attr.adb
	(Build_Array_VS_Func and Build_Record_VS_Func): Pass in the
	Attr as the Related_Node parametr when calling
	Make_Temporary for the Func_Id for the array and record
	Valid_Scalars local functions.