ada: Warning for composite equality that calls an abstract equality function
When equality is tested for a composite type that has any record
components whose type has an abstract equality function that will
be called as part of the enclosing type's equality, Program_Error
will be raised.  We now issue a warning on the equality test,
mentioning the component type whose abstract equality function
will trigger the exception.  Note that this is currently only
done for top-level components of the composite type.  Another
limitation is that the warning is not issued when the outer
composite type is tagged.

gcc/ada/ChangeLog:

	* exp_ch4.adb (Expand_N_Op_Eq): Check for warning about call to
	the abstract equality function of a component type, for both array
	and record enclosing types.
	(Warn_On_Abstract_Equality_For_Component): New procedure to issue
	a warning when an abstract equality function of a component type
	will be called and result in Program_Error.