ada: Refine condition for reporting warnings on components with abstract equality
The initial implementation of the warning resulted in unwanted false
positives for types that have a user-defined equality function (in
which case abstract equality on components will typically not ever
be invoked).  The conditions for reporting the warning are refined
by this change to exclude checking for presence of abstract component
equality functions in the case where the containing type has a user-defined
equality.

gcc/ada/ChangeLog:

	* exp_ch4.adb (Expand_N_Op_Eq): Test for absence of user-defined
	equality on type being compared (for both array and record types)
	as a condition for checking for abstract equality on component
	types. Add a "???" comment about current limitations on issuing
	the new warning.
	(Warn_On_Abstract_Equality_For_Component): Remove temporary disabling
	of the warning. Improve comment on declaration.