ada: Apply ghost regions for assigmnents correctly
When frontend is operating in GNATprove mode (where expander is disabled), it
should check ghost policy for assignment statements just like it does for other
statements. This is because we want ghost policy errors to be reported not just
by GNAT, but also by GNATprove.

Additionally we need to perform the checks for valid location of ghost assigments
based on the region around the assigment before we create the region for
the assignment itself.

gcc/ada/ChangeLog:

	* ghost.adb (Mark_And_Set_Ghost_Assignment): Create a ghost region
	for an assigment irregardless of whether the expander is active.
	Relocate the Assignment validity checks from Is_OK_Statement to
	this subprogram.