ada: Avoid ghost context check in early freeze
We freeze the expression for expression functions that complete
before we have set up the ghost region from both the existing spec
and body. Avoid triggering the ghost context checks during the
analysis for this early freeze as the expression will be reanalyzed
when we analyze the new function body created for the expression
function.

gcc/ada/ChangeLog:

	* sem_ch6.adb (Analyze_Expression_Function): Disable ghost
	checks during the early freeze.