ada: Refactor ghost argument consistency checks
Create a new method for checking and emitting errors on pragmas
Unused, Unrefefrenced, Unreferenced_Objects, Inline and No_Return
that support specifying multiple entities as arguments.

Emit an error when one argument is ghost and the other is not and
when one argument has a ghost policy check and the other has ghost
policy ignore.

Update the Suppressed_Ghost_Policy_Check_Pragma list pragma Inline
that should be there to avoid an incorrect invalid pragma context
error.

gcc/ada/ChangeLog:

	* sem_prag.adb (Check_Inconsistent_Argument_Ghostliness):
	new method for handling the ghost constency errors between
	different arguments. Use this method in the processing for
	pragmas Unused, Unrefefrenced, Unreferenced_Objects, Inline and
	No_Return.
	* sem_prag.ads (Suppressed_Ghost_Policy_Check_Pragma): add
	missing entry for pragma Inline.