ada: Incorrect expansion of VADS Size attribute applied to an array type
The compiler fails, either with an Assert_Failure (for a compiler with
assertions enabled) or by producing a bad value, when processing a VADS_Size
attribute (or a Size attribute when pragma Use_VADS_Size is in force),
in the case where the prefix denotes a composite type or subtype whose
object size (Esise) is not known by the front end. The bug is due to
Eval_Attribute attempting to fold a VADS-related size attribute even
when Esize is not yet known for the type.

gcc/ada/ChangeLog:

	* sem_attr.adb (Eval_Attribute): In the case of a VADS-specific
	size attribute, only fold to the object size when the Esize field
	is known.