Fix -Wlto-type-mismatch warning during GNAT LTO build
The recent addition of Pragma_Unsigned_Base_Range to the enumeration type
Pragma_Id has made it exceed 256 enumeration values and thus overflow the
unsigned 8-bit type used for it on the C side.  This should have stopped
the build, except that a glitch in the Snames machinery causes one value
to be dropped on the C side, leaving it at just 256 enumeration values.

This fixes both issues, i.e. ensures that the number of enumeration values
is the same on both sides and bumps the size of the C type to 16 bits.

gcc/ada/
	PR ada/121885
	* snames.ads-tmpl (Pragma_Id): Rename Unknown_Pragma to
	Pragma_Unknown for the sake of XSnamesT.
	* snames.adb-tmpl (Get_Pragma_Id): Adjust to above renaming.
	* snames.h-tmpl (Attribute_Id): Change underlying type to Byte.
	(Get_Attribute_Id): Use Byte as return value.
	(Pragma_Id): Change underlying type to Word.
	(Get_Pragma_Id): Use Word as return value.
	* types.h (Word): New typedef.
	* exp_prag.adb (Expand_N_Pragma): Remove useless comment.
	* par-prag.adb (Prag): Adjust to above renaming.
	* sem_prag.adb (Analyze_Pragma): Likewise.
	(Sig_Flags): Likewise.