ada: Add Visitor generic to Repinfo
For a gnat-llvm debuginfo patch, it was convenient to be able to
inspect the expressions created during back-annotation.  This patch
adds a new generic Visit procedure that can be implemented to allow
such inspection.  List_GCC_Expression is reimplemented in terms of
this procedure as a proof of concept.

gcc/ada/ChangeLog:

	* repinfo.adb (Visit): New procedure.
	(List_GCC_Expression): Rewrite.
	* repinfo.ads (Visit): New generic procedure.