ada: Add new abstractions to Table.Table
This patch adds two new subprograms to Table.Table: Clear and Is_Empty.
Their selling point is that they don't require being aware of the bounds
of the instance of Table.Table, avoiding the off-by-one errors that can
happen when using Set_Last or Last directly.

This patch also replaces existing code by calls to these new subprograms
in a few places where it makes sense. It also adds a call to
Table.Table.First in the same spirit on the side.

gcc/ada/ChangeLog:

	* table.ads (Clear, Is_Empty): New subprograms.
	* table.adb (Clear, Is_Empty): Likewise.
	(Init): Use new subprogram.
	* atree.adb (Traverse_Func_With_Parent): Use new subprograms.
	* fmap.adb (Empty_Tables): Use new subprogram.
	* par_sco.adb (Process_Pending_Decisions): Likewise.
	* sem_elab.adb (Check_Elab_Call): Likewise.
	* sem_ch12.adb (Build_Local_Package, Analyze_Package_Instantiation,
	Analyze_Subprogram_Instantiation): Likewise.
	(Save_And_Reset): Use Table.Table.First.