ada: Corrupted unwind info in aarch64-vx7r2 llvm kernel tests
Adjust the register restoration on aarch64 to not use register 96
on llvm. Avoids the "reg too big" warning on aarch64 when sigtramp
is called.  For llvm and aarch64, the correct choice seems to be 32.

Remove parens on REGNO_PC_OFFSET when compiling,
it causes a silent failure due to alphanumeric register names.

Define a macro for __attribute ((optimize (2))) which is
empty if not availble. (Despite being documented, it generates an
"unknown attribute" warning with clang.)

Define ATTRIBUTE_PRINTF_2 if not defined.

gcc/ada/ChangeLog:

	* sigtramp-vxworks-target.h (REGNO_PC_OFFSET): Use 32 vice
	96 with llvm/clang. (REGNO_G_REG_OFFSET): Remove parens on
	operand. (REGNO_GR): Likewise.
	* sigtramp-vxworks.c (__gnat_sigtramp): Define a macro for
	__attribute__ optimize, which is empty of not available.
	* raise-gcc.c (db): Define ATTRIBUTE_PRINTF_2 if not defined.