ada: Get rid of Sy/Sm mixing (Chars)
We should not mix "syntactic" and "semantic" for the same field
in different node kinds.

The Chars field is both syntactic and semantic. This patch
makes it always syntactic, and does some other Chars-related
cleanups.

An attempt was made to instead rename the semantic field
to be Op_Chars, but that complicates things, because there
is a fair amount of code that fetches the Chars field
without knowing the node kind. Notably, Errout does this.

No change in overall compiler behavior.

gcc/ada/ChangeLog:

	* gen_il-gen-gen_nodes.adb (N_Op):
	Make Chars syntactic, and move it down into subclasses
	N_Binary_Op and N_Unary_Op.
	* gen_il-gen.adb (Create_Type):
	Do not exempt Chars from the ordering rule.
	(Exception_To_Inheritance_Rule): Exempt Chars from the
	inheritance rule.
	(Check_For_Syntactic_Field_Mismatch):
	Do not exempt Chars from the syntactic mismatch rule.
	This is the main point of this change.
	(Put_Make_Bodies): The Nmake functions for types in N_Op
	will now take a Chars parameter, which should always
	default to No_Name. This will be overwritten by the
	special-case Set_Chars call. Assert that it is in
	fact defaulted.
	* exp_ch4.adb (Expand_Array_Comparison):
	Use the Nkind instead of the Chars, which seems cleaner.
	Use a case instead of an elsif chain.
	* sem_attr.adb (Proper_Op): Minor cleanup.
	* sem_ch8.adb: Minor reformatting.
	* sem_res.adb (Operator_Kind): Tighten up the result subtype.