ada: Fix visibility issue in nested instance with use clause for formal package
The compiler gives a spurious visibility error for a formal object parameter
of a formal package with a use clause, present in a parent instance, when an
instance of a child generic unit is compiled, while this does not happen in
the same circumstances for a formal type parameter.

The discrepancy comes from the Check_Generic_Actuals procedure, which clears
the Is_Hidden flag on all the actuals of a generic instance, but only sets
the Is_Potentially_Use_Visible flag on the actuals for formal types.

The change also contains a fix for a pasto in Restore_Private_Views, which
is responsible for undoing the changes made by Check_Generic_Actuals.

gcc/ada/ChangeLog:

	PR ada/122161
	* sem_ch12.adb (Check_Generic_Actuals): Consistently set the
	Is_Potentially_Use_Visible flag on actuals whenever the Is_Hidden
	flag is cleared.
	(Restore_Private_Views): Clear the Is_Potentially_Use_Visible flag
	explicitly on the entities of an actual package.