ada: Get rid of Sy/Sm mixing (Default_Expression)
We should not mix "syntactic" and "semantic" for the same field
in different node kinds.

Change the name of Default_Expression (a syntactic field) on
N_Formal_Object_Declaration to be Expression. This avoids
the conflict with the name-named semantic field of
N_Parameter_Specification.

It is also more uniform with other syntactic categories that
use default_expression in the syntax rules.  See, for example,
COMPONENT_DECLARATION in sinfo.ads, which is presumably intended
to be uniform with OBJECT_DECLARATION.

Fix the comment on N_Formal_Object_Declaration to refer
to the new field name Expression.

gcc/ada/ChangeLog:

	* gen_il-gen-gen_nodes.adb: Rename Default_Expression to be
	Expression.
	* gen_il-gen.adb (Check_For_Syntactic_Field_Mismatch): Do not
	exempt Default_Expression from the rule.
	* par-ch12.adb (P_Formal_Object_Declarations): Use renamed setter.
	* sem_ch12.adb (Default): Use renamed getter.
	(Analyze_Formal_Object_Declaration): Likewise.
	* sprint.adb (Sprint_Node_Actual): Use renamed getter.
	* sinfo.ads: Fix comments.