ada: Fix incorrect static string concatenation with null left string
It comes from the implementation of an optimization for static concatenation
in Resolve_String_Literal, which causes the original subtype of the literal
to be lost.  Now this subtype must be preserved in the case where the left
operand of the concatenation may be null, per the 4.5.3(5) subclause.

gcc/ada/ChangeLog:

	PR ada/122160
	* sem_res.adb (Resolve_Op_Concat_Rest): Do not build the subtype of
	the second operand again if it has already been built.
	(Resolve_String_Literal): Do not defer the creation of the subtype
	for the right operand of a concatenation whose left operand may be
	the null string.