ada: Avoid duplicate streaming and Put_Image subprograms.
Duplicate streaming and Put_Image subprograms were being generated in some
cases where this was not intended. In most cases this only resulted in unwanted
code duplication (which, of course, is not good), but in some cases it resulted
in compilation failures with spurious "duplicate body" error messages.

gcc/ada/ChangeLog:

	* exp_attr.adb: Rewrite the spec and implementation of package
	Cached_Attribute_Ops so that the saved value associated with a
	type in a given map is not a single subprogram but instead a
	set of subprograms. Thus, the correct generation of a second subprogram
	for given type for use in some other context no longer causes the
	first subprogram to be forgotten. This allows more reuse and,
	in particular, allows reuse in the case where generating another
	copy of the subprogram would result in a compilation failure.
	Update Cached_Attribute_Ops clients correspondingly.